python_apps/agent_studio/pyproject.toml
[project]
name = "agent-studio"
version = "0.1.0"
description = "Agent Studio"
requires-python = ">=3.11"
dependencies = [
    # Core SDK - provides workflow engine, tools, database models
    "workflow-core-sdk",
    # FastAPI and web
    "fastapi>=0.111.0",
    "uvicorn>=0.34.2",
    "python-multipart",
    # Legacy dependencies (to be phased out)
    "bs4>=0.0.2",
    "httpx>=0.27.0",
    "markdownify>=1.1.0",
    "openai>=1.77.0",
    "pydantic>=2.7.1",
    "python-dotenv>=1.0.1",
    "requests>=2.31.0",
    "sqlalchemy>=2.0.30",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.1",
    "pika>=1.3.2",
    "redis>=4.5.1",
    "elevaite-logger",
    "pytest-asyncio>=0.26.0",
    "google-auth>=2.40.0",
    "google-api-python-client>=2.0.0",
    "qdrant-client>=1.7.0",
    "simple-salesforce>=1.12.6",
    "aiofiles>=23.2.0",
    "elevaite-ingestion",
    "sentence-transformers>=5.1.0",
    "scikit-learn>=1.6.1",
    "py2pdf>=0.1.3",
    "pypdf2>=3.0.1",
    "boto3>=1.34.0",
    "opentelemetry-instrumentation-fastapi>=0.57b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.57b0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0.0", "pytest-asyncio>=0.21.0", "pytest-cov>=4.1.0"]

[tool.uv.sources]
elevaite-logger = { workspace = true }
elevaite-ingestion = { workspace = true }
workflow-core-sdk = { workspace = true }
